move input checking method to util
